requirements.txt
python-dotenv>=1.0.0
psycopg2-binary>=2.9
httpx>=0.27
fastapi>=0.115
uvicorn>=0.32
websockets>=13.0
numpy>=1.26
scikit-learn>=1.4
python-multipart>=0.0.7
